feat(group) : 그룹 예외 처리 추가

+++ src/main/java/com/homelearn/back/group/GroupServiceImpl.java
@@ -2,9 +2,14 @@ package com.homelearn.back.group;
 
 import com.homelearn.back.group.dto.*;
 import com.homelearn.back.group.entity.Group;
+import com.homelearn.back.group.entity.GroupItem;
+import com.homelearn.back.group.exception.GroupErrorCode;
+import com.homelearn.back.group.exception.GroupException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
+
+import static com.homelearn.back.group.exception.GroupErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -20,21 +25,33 @@ public class GroupServiceImpl implements GroupService{
 
     @Override
     public void deleteGroup(Long groupId) {
+        Group group = groupMapper.findGroup(groupId);
+        if (group!=null){
+            throw new GroupException(NOT_EXISTS_GROUP);
+        }
         groupMapper.deleteGroup(groupId);
     }
 
     @Override
-    public List<Group> findGroupListByUserId(Long groupId) {
-        return groupMapper.findGroupListByUserId(groupId);
+    public List<Group> findGroupListByUserId(Long userId) {
+        return groupMapper.findGroupListByUserId(userId);
     }
 
     @Override
     public void addGroupItem(GroupItemInputSpec groupItemInputSpec) {
+        GroupItem groupItem = groupMapper.findGroupItem(groupItemInputSpec);
+        if (groupItem!=null){
+            throw new GroupException(ALREADY_IN_GROUP_ITEM);
+        }
         groupMapper.addGroupItem(groupItemInputSpec);
     }
 
     @Override
     public void deleteGroupItem(GroupItemInputSpec groupItemInputSpec) {
+        GroupItem groupItem = groupMapper.findGroupItem(groupItemInputSpec);
+        if (groupItem==null){
+            throw new GroupException(NOT_EXISTS_GROUP_ITEM);
+        }
         groupMapper.deleteGroupItem(groupItemInputSpec);
     }
